Split GitVersionControlSpec into interface and implementation
It is just healthy to keep these separate.

Part of gradle/gradle-native#88
