+++ subprojects/version-control/src/main/java/org/gradle/vcs/git/GitVersionControlSpec.java
@@ -17,19 +17,15 @@
 package org.gradle.vcs.git;
 
 import org.gradle.api.Incubating;
-import org.gradle.internal.UncheckedException;
 import org.gradle.vcs.VersionControlSpec;
 
 import java.net.URI;
-import java.net.URISyntaxException;
 
 /**
  * A specification of a Git repository.
  */
 @Incubating
-public class GitVersionControlSpec implements VersionControlSpec {
-    private URI url;
-
+public interface GitVersionControlSpec extends VersionControlSpec {
     /**
      * The URL for the repository in the specification.
      *
@@ -39,45 +35,15 @@ public class GitVersionControlSpec implements VersionControlSpec {
      * operations which are both unsuited for simple data specification and
      * allocate additional memory.</p>
      */
-    public URI getUrl() {
-        return url;
-    }
+    URI getUrl();
 
     /**
      * Sets the URL of the repository.
      */
-    public void setUrl(URI url) {
-        this.url = url;
-    }
+    void setUrl(URI url);
 
     /**
      * Sets the URL of the repository.
      */
-    public void setUrl(String url) {
-        try {
-            setUrl(new URI(url));
-        } catch (URISyntaxException e) {
-            throw new UncheckedException(e);
-        }
-    }
-
-    @Override
-    public String getDisplayName() {
-        return "Git Repository at " + getUrl();
-    }
-
-    @Override
-    public String getRepositoryId() {
-        String host = url.getHost();
-        if (host == null) {
-            host = "local";
-        }
-        return host + "/" + url.getPath();
-    }
-
-    @Override
-    public String getRepoName() {
-        String[] pathParts = url.getPath().split("/");
-        return pathParts[pathParts.length-1];
-    }
+    void setUrl(String url);
 }
